Change htif_t API, use getopt for option parsing
This changes the `htif_t` constructor to use getopt for option parsing
adding support for POSIX-like `--` long options while preserving
support for "legacy" `+` args. The options supported by `htif_t` are
exposed to programs importing the `htif.h` header.

This results in a change of the `htif_t` and `dtm_t` constructor APIs
so that it can directly process command line arguments. Previously
the `htif_t` constructor was:
  htif_t::htif_t(const std::vector<std::string>& target_args)
Now it is:
  htif_t::htif_t(int argc, char** argv)

As `htif_t` is a secondary consumer of command line arguments (e.g.,
it is constructed after the rocket-chip emulator), the options
supported by `htif_t` are exposed via three new macros in `htif_t.h`:
  * HTIF_USAGE_OPTIONS -- usage string suitable to be appended to any
    program that will pass arguments to `htif_t`
  * HTIF_LONG_OPTIONS_OPTIND -- Base val for `htif_t` options
  * HTIF_LONG_OPTIONS -- `struct option` fragment for `htif_t`. All
    options index starting from `HTIF_LONG_OPTIONS_OPTIND`.

Support for the currently unsupported option `+disk` now raises an
exception.

Space is provided for non-standard options which may be exposed to
`htif_t`. Currently only the VCS options of the form `-cm XXX` are
ignored.