Cleanup, add argumentless htif_t constructor
Does some minor cleanup of usage formatting.

Adds a htif_t() constructor that does not do argument parsing. This is
needed for something like `elf2hex` which wants to pass no arguments
to htif_t.

Note: `elf2hex` is currently non-working on master and not as a result
of this.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@gmail.com>